+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add guide for content sharing
 - Display meeting id in the demo app
+- Add additional callback in AudioVideoObserver to indicate video downlink pressure
 
 ### Changed
 - Prevent prebuild from increase patch number when publishing to NPM
@@ -18,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - Remove unimplemented callbacks remoteDidMuteAudio and remoteDidUnmuteAudio on AudioVideoObserver
+
+
 
 ## [1.2.1] - 2020-03-20